check if *.patch exists before removing to suppress outspam

+++ cherry-picks.sh
@@ -12,7 +12,10 @@
 # ---------------------------------------------------------
 
 function check_clean {
-  rm *.patch
+  if [ -d "*.patch" ]
+  then
+    rm *.patch
+  fi
   if [ -d ".git/rebase-apply" ]
   then
     git am --abort
